moved logger around
