+++ pkg/azure/keyvault/init.go
@@ -9,12 +9,11 @@ import (
 var vaultName, setDebug, convertHyphenToUnderscores string
 
 func init() {
-	if os.Getenv("AZURE_TENANT_ID") == "" || os.Getenv("AZURE_CLIENT_ID") == "" || os.Getenv("AZURE_CLIENT_SECRET") == "" || os.Getenv("KVAULT") == "" {
+	if os.Getenv("AZURE_TENANT_ID") == "" || os.Getenv("AZURE_CLIENT_ID") == "" || os.Getenv("AZURE_CLIENT_SECRET") == "" || os.Getenv("VAULT_NAME") == "" {
 		fmt.Println("env vars not set, exiting...")
 		os.Exit(1)
 	}
 
-	vaultName = os.Getenv("KVAULT")
-	setDebug = strings.ToLower(os.Getenv("DEBUG"))
+	vaultName = os.Getenv("VAULT_NAME")
 	convertHyphenToUnderscores = strings.ToLower(os.Getenv("CONVERT_HYPHENS_TO_UNDERSCORES"))
 }
